+++ CHANGELOG.md
@@ -88,7 +88,6 @@ Candidate release for the "all clear" milestone. There wasn't any real usage of 
   - ++`RToken.IssuanceThrottleSet`
   - ++`RToken.RedemptionThrottleSet`
 - Allow redemption while DISABLED
-- Disallow staking while FROZEN
 - Allow `grantRTokenAllowances()` while paused
 - Add `RToken.monetizeDonations()` escape hatch for accidentally donated tokens
 - Collateral default threshold: 5% -> 1% (+ include oracleError)
@@ -101,4 +100,5 @@ Candidate release for the "all clear" milestone. There wasn't any real usage of 
 - Payout RSR rewards on `StRSR.setRatio()`
 - Distinguish oracle timeouts when dealing with multiple oracles in one plugin
 - Add safety during asset degregistration to ensure it is always possible to unregister an infinite-looping asset
--
+- Fix `StRSR`/`RToken` EIP712 typehash to use release version instead of "1"
+- Add `FacadeRead.redeem(IRToken rToken, uint256 amount, uint48 basketNonce)` to return the expected redemption quantities on the basketNonce, or revert